Move the display logic of containers to a CLI helper
